[#1189] Reorganize base view helpers and serializers
* Extracts i18n functionality into dedicated I18nHelpers module
* Moves JS variable handling to new JSVarsHelpers module
* Renames view helper files to better reflect their purpose
* Consolidates serializer modules in dedicated directory

The refactoring improves code maintainability by separating concerns
into focused, single-responsibility modules.
