+++ apps/web/core/views/serializers/domain_serializer.rb
@@ -1,6 +1,4 @@
-
-
-
+# apps/web/core/views/serializers/domain_serializer.rb
 
 module Core
   module Views
