+++ apps/web/core/views/serializers/htmltags_serializer.rb
@@ -1,0 +1,9 @@
+# apps/web/core/views/serializers/htmltags_serializer.rb
+
+module Core
+  module Views
+    module HTMLTags
+
+    end
+  end
+end
