--- apps/web/core/views/serializers/metatags.rb
@@ -1,9 +1,0 @@
-
-
-module Core
-  module Views
-    module MetaTags
-
-    end
-  end
-end
