+++ apps/web/core/views/serializers/plan_serializer.rb
@@ -1,5 +1,4 @@
-
-
+# apps/web/core/views/serializers/plan_serializer.rb
 
 module Core
   module Views
