+++ apps/web/core/views/serializers/template_serializer.rb
@@ -1,4 +1,4 @@
-
+# apps/web/core/views/serializers/template_serializer.rb
 
 module Core
   module Views
