+++ tests/unit/ruby/try/11_cached_method_try.rb
@@ -27,7 +27,7 @@ OT.boot! :test, false
 @num = rand(1000)
 
 class TestHelper
-  include Core::ViewHelpers
+  include Core::Views::JSVarsHelpers
   attr_reader :num
 
   def initialize(num)
